Move Reshuffle and ReshuffleTrigger out of util
These are two deprecated, but regularly used, transforms and triggers, to the
point where the Runner API proto even includes something akin to the
ReshuffleTrigger.

They are moved to where they "belong" and marked deprecated.
