+++ runners/spark/src/main/java/org/apache/beam/runners/spark/translation/GroupCombineFunctions.java
@@ -24,6 +24,7 @@ import org.apache.beam.runners.spark.util.ByteArray;
 import org.apache.beam.sdk.coders.Coder;
 import org.apache.beam.sdk.coders.IterableCoder;
 import org.apache.beam.sdk.coders.KvCoder;
+import org.apache.beam.sdk.transforms.Reshuffle;
 import org.apache.beam.sdk.util.WindowedValue;
 import org.apache.beam.sdk.util.WindowedValue.WindowedValueCoder;
 import org.apache.beam.sdk.values.KV;
@@ -223,7 +224,7 @@ public class GroupCombineFunctions {
 
   /**
    * An implementation of
-   * {@link org.apache.beam.sdk.util.Reshuffle} for the Spark runner.
+   * {@link Reshuffle} for the Spark runner.
    */
   public static <K, V> JavaRDD<WindowedValue<KV<K, V>>> reshuffle(
       JavaRDD<WindowedValue<KV<K, V>>> rdd,
