+++ sdks/java/core/src/main/java/org/apache/beam/sdk/transforms/ReifyTimestamps.java
@@ -16,11 +16,8 @@
  * limitations under the License.
  */
 
-package org.apache.beam.sdk.util;
+package org.apache.beam.sdk.transforms;
 
-import org.apache.beam.sdk.transforms.DoFn;
-import org.apache.beam.sdk.transforms.PTransform;
-import org.apache.beam.sdk.transforms.ParDo;
 import org.apache.beam.sdk.values.KV;
 import org.apache.beam.sdk.values.PCollection;
 import org.apache.beam.sdk.values.TimestampedValue;
@@ -29,7 +26,7 @@ import org.apache.beam.sdk.values.TimestampedValue;
  * {@link PTransform PTransforms} for reifying the timestamp of values and reemitting the original
  * value with the original timestamp.
  */
-public class ReifyTimestamps {
+class ReifyTimestamps {
   private ReifyTimestamps() {}
 
   /**
