+++ sdks/java/core/src/main/java/org/apache/beam/sdk/transforms/Reshuffle.java
@@ -15,15 +15,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.beam.sdk.util;
+package org.apache.beam.sdk.transforms;
 
-import org.apache.beam.sdk.transforms.DoFn;
-import org.apache.beam.sdk.transforms.GroupByKey;
-import org.apache.beam.sdk.transforms.PTransform;
-import org.apache.beam.sdk.transforms.ParDo;
+import org.apache.beam.sdk.annotations.Internal;
 import org.apache.beam.sdk.transforms.windowing.BoundedWindow;
+import org.apache.beam.sdk.transforms.windowing.ReshuffleTrigger;
 import org.apache.beam.sdk.transforms.windowing.TimestampCombiner;
 import org.apache.beam.sdk.transforms.windowing.Window;
+import org.apache.beam.sdk.util.IdentityWindowFn;
 import org.apache.beam.sdk.values.KV;
 import org.apache.beam.sdk.values.PCollection;
 import org.apache.beam.sdk.values.TimestampedValue;
@@ -31,18 +30,22 @@ import org.apache.beam.sdk.values.WindowingStrategy;
 import org.joda.time.Duration;
 
 /**
- * A {@link PTransform} that returns a {@link PCollection} equivalent to its input but operationally
- * provides some of the side effects of a {@link GroupByKey}, in particular preventing fusion of
- * the surrounding transforms, checkpointing and deduplication by id (see
- * {@link ValueWithRecordId}).
+ * <b>For internal use only; no backwards compatibility guarantees.</b>
  *
- * <p>Performs a {@link GroupByKey} so that the data is key-partitioned. Configures the
- * {@link WindowingStrategy} so that no data is dropped, but doesn't affect the need for
- * the user to specify allowed lateness and accumulation mode before a user-inserted GroupByKey.
+ * <p>A {@link PTransform} that returns a {@link PCollection} equivalent to its input but
+ * operationally provides some of the side effects of a {@link GroupByKey}, in particular preventing
+ * fusion of the surrounding transforms, checkpointing and deduplication by id.
+ *
+ * <p>Performs a {@link GroupByKey} so that the data is key-partitioned. Configures the {@link
+ * WindowingStrategy} so that no data is dropped, but doesn't affect the need for the user to
+ * specify allowed lateness and accumulation mode before a user-inserted GroupByKey.
  *
  * @param <K> The type of key being reshuffled on.
  * @param <V> The type of value being reshuffled.
+ * @deprecated this transform's intended side effects are not portable; it will likely be removed
  */
+@Internal
+@Deprecated
 public class Reshuffle<K, V> extends PTransform<PCollection<KV<K, V>>, PCollection<KV<K, V>>> {
 
   private Reshuffle() {
