+++ sdks/java/io/kafka/src/main/java/org/apache/beam/sdk/io/kafka/serialization/InstantSerializer.java
@@ -19,7 +19,6 @@
 package org.apache.beam.sdk.io.kafka.serialization;
 
 import java.util.Map;
-
 import org.apache.kafka.common.serialization.LongSerializer;
 import org.apache.kafka.common.serialization.Serializer;
 import org.joda.time.Instant;
